Refactor sharding sphere lock manager
